+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 * Refactored `Connection` interface
 * Removed `CustomOption` and taking client with custom options
 * Removed `proxy` package
-* Improved `db.With()` helper for create child connections
-* Setted shared `conn.Pool` for all child `ydb.Connection`  
+* Improved `db.With()` helper for child connections creation
+* Set shared `conn.Pool` for all children `ydb.Connection`  
 
 ## 3.12.1
 * Added `trace.Driver.OnConnPark` event